Fixed Players energy management system to handle more than one building at a time
More than one building would clog the list with copies of requests. So I allow only one of each request to be placed. This caused an issue where a copy of the request would be left inside the list of requests, but the energy requested int wouldn't reflect that. This happened when it finished building, making it so that no further objects could be built properly. I've got a hotfix but I'm not certain it is reliable and won't cause further errors.